==> notebooks/01_load_and_validate_raw_data.ipynb <==
import os
from pathlib import Path
import requests


def download_one_file_of_raw_data(year: int, month: int) -> Path:
    """"""
    URL = f"https://d37ci6vzurychx.cloudfront.net/trip-data/yellow_tripdata_{year}-{month:02d}.parquet"
    response = requests.get(URL)

    if response.status_code == 200:
        path = f'../data/raw'
        if not os.path.exists(path):
            os.makedirs(path)
        file_path = f'{path}/rides_{year}-{month:02d}.parquet'
        open(file_path, 'wb').write(response.content)
        return file_path
    else:
        raise Exception(f'{URL} is not available.')


year=2022
month=1
download_one_file_of_raw_data(year=year, month=month)

import pandas as pd

rides = pd.read_parquet(f'../data/raw/rides_{year}-{month:02d}.parquet')
rides.head()

rides = rides[['tpep_pickup_datetime', 'PULocationID']]
rides.rename(columns={
    'tpep_pickup_datetime': 'pickup_datetime', 
    'PULocationID': 'pickup_location_id'
}, inplace=True)

rides.head()

# (7:18')Step 1. Fetch raw data and validate it
rides['pickup_datetime'].describe()

rides = rides[rides['pickup_datetime'] >= '2022-01-01']
rides = rides[rides['pickup_datetime'] < '2022-02-01']
rides['pickup_datetime'].describe()

path = f'../data/transformed'
if not os.path.exists(path):
    os.makedirs(path)
file_path = f'{path}/validated_rides_{year}-{month:02d}.parquet'
rides.to_parquet(file_path)
