==> notebooks/02_transform_raw_data_into_ts_data.ipynb <==
# Transforming time series data into tabular data for training machine learning models 

involves creating features that capture relevant patterns and information from the time series. Here are some common strategies:

1. **Temporal Aggregation:**
   - Aggregate the taxi rides data based on time intervals, such as hourly, daily, or weekly.
   - Calculate summary statistics for each time interval, e.g., total rides, average rides, maximum rides, etc.

2. **Time Lag Features:**
   - Create lag features that represent past observations. For example, use the number of rides in the previous hour or previous day as features.
   - Lag features can capture trends and seasonality in the data.

3. **Rolling Window Statistics:**
   - Calculate rolling window statistics, such as rolling mean or rolling standard deviation, to smooth out short-term fluctuations and capture longer-term trends.

4. **Time Decomposition:**
   - Decompose the time series into trend, seasonality, and residual components using methods like seasonal decomposition of time series (STL) or moving averages.
   - Use the decomposed components as features.

5. **Date and Time Features:**
   - Extract date and time-related features, such as day of the week, month, hour, and whether it's a holiday or not.
   - These features can help capture day-of-week effects, monthly trends, and other temporal patterns.

6. **Historical Aggregations:**
   - Calculate historical aggregations over different time spans, such as the average number of rides over the past week or month.
   - These features can capture longer-term trends and patterns.

7. **Windowed Statistics:**
   - Compute statistics over fixed-size windows of time, such as the median rides in the last 3 hours or the minimum rides in the last 24 hours.

8. **Time Since Last Event:**
   - Create features based on the time elapsed since certain events, such as the last peak in rides or the last significant drop.

9. **Weather or External Factors:**
   - Integrate external factors like weather conditions, holidays, or special events that might influence taxi ride demand.

10. **Encoding Cyclic Features:**
    - Encode cyclic features like time of day or day of the week using techniques such as circular encoding to preserve the cyclical nature of these features.

11. **Slice and Slide**
    - Strategy that involves creating overlapping windows or slices of the time series data and using these slices as input features for the model. This approach is particularly useful for capturing temporal patterns and dependencies in the data. Here's a brief overview:

   - **Slice**:

      Divide the time series data into fixed-size slices or windows. Each slice covers a specific period, and these slices can overlap or be non-overlapping.
      For example, you could create slices of one-hour intervals, capturing the taxi rides data for each hour.

   - **Slide**:

      Move the window through the time series with a specified step size. This step size determines how much the window advances at each iteration.
      Overlapping slices allow the model to learn from multiple perspectives of the time series.
      Features from Slices:

      Extract relevant features from each slice, such as the total number of rides, average rides, or other summary statistics within that time window.
      These features become the input variables for training the machine learning model.
      Target Variable:

      Define the target variable for prediction, typically associated with the next time point after the end of each slice.
      For example, if using one-hour slices, the target variable might be the number of rides in the next hour.
      Model Training:

      Train the machine learning model using the extracted features from the slices and the corresponding target variable.
      The model learns to capture temporal dependencies and patterns within each slice.
      This "Slice and Slide" strategy is especially effective when there is a need to model short-term dependencies and fluctuations in the time series data. It enables the model to consider the recent history and adapt to changing patterns over time. 
      
      However, the choice of window size and step size is crucial and should be determined based on the characteristics of the data and the specific problem being addressed. Too short a window might miss important patterns, while too long a window might smooth out relevant information. It's often necessary to experiment with different configurations to find the optimal parameters for the problem at hand.


import pandas as pd

pd.set_option('display.max_columns', None)


rides = pd.read_parquet('../data/transformed/validated_rides_2022-01.parquet')
rides.head()

# Each row represent a ride
""">>CHECKLIST: decide which frequency to work on. We are going to work with a time series data at an hour frequency.
The floor function basically takes the pickup data column and round it to the closest previous hour<<"""
rides['pickup_hour'] = rides['pickup_datetime'].dt.floor('H')
rides

rides.info()

# get nr rows in each group
agg_rides = rides.groupby(['pickup_hour', 'pickup_location_id']).size().reset_index()
agg_rides.rename(columns={0: 'rides'}, inplace=True)
agg_rides

""">>CHECKLIST: Step 2. Transform raw validated data into time-series data
Pickup location with low demand has sparse series, so some hours might have zero rides and they dont appear in the training ts data.
We want to add missing slots in the data with zero rides (target).<<"""

from tqdm import tqdm

def add_missing_slots(agg_rides: pd.DataFrame) -> pd.DataFrame:
    """Create a complete time series range of equally spaced time points (where the difference 
    between any two adjacent points is specified by the given frequency).
    
    This operation is performed per location_id"""
    
    location_ids = agg_rides['pickup_location_id'].unique()
    full_range = pd.date_range(
        agg_rides['pickup_hour'].min(), agg_rides['pickup_hour'].max(), freq='H')
    output = pd.DataFrame()
    for location_id in tqdm(location_ids):

        # keep only rides for this 'location_id'
        agg_rides_i = agg_rides.loc[agg_rides.pickup_location_id == location_id, ['pickup_hour', 'rides']]
            
        # quick way to add missing dates with 0 in a Series
        # taken from https://stackoverflow.com/a/19324591
        agg_rides_i.set_index('pickup_hour', inplace=True)
        agg_rides_i.index = pd.DatetimeIndex(agg_rides_i.index)
        agg_rides_i = agg_rides_i.reindex(full_range, fill_value=0)
        
        # add back `location_id` columns
        agg_rides_i['pickup_location_id'] = location_id

        output = pd.concat([output, agg_rides_i])
    
    # move the purchase_day from the index to a dataframe column
    output = output.reset_index().rename(columns={'index': 'pickup_hour'})
    
    return output

agg_rides.shape

agg_rides_all_slots = add_missing_slots(agg_rides)

agg_rides_all_slots.shape

""">>CHECKLIST: (0:00' ) Plot the time-series data
visualize data to detect bugs in data processing<<"""
from typing import Optional, List
import plotly.express as px

# TODO: add a orange faded line in every frequency defined as a parameter. for example every week or month

def plot_rides(
    rides: pd.DataFrame,
    locations: Optional[List[int]] = None
    ):
    """
    Plot time-series data
    """
    rides_to_plot = rides[rides.pickup_location_id.isin(locations)] if locations else rides

    fig = px.line(
        rides_to_plot,
        x="pickup_hour",
        y="rides",
        color='pickup_location_id',
        template='none',
    )

    fig.show()

plot_rides(agg_rides_all_slots, locations=[43])

agg_rides_all_slots.to_parquet('../data/transformed/ts_data_2022_01.parquet')
